+++ src/main/java/baseball/control/ComputerControl.java
@@ -1,8 +1,6 @@
 package baseball.control;
 
-import static baseball.model.NumConstant.MAX_NUMBER;
-import static baseball.model.NumConstant.MIN_NUMBER;
-import static baseball.model.NumConstant.NUM_LENGTH;
+import static baseball.constant.NumConstant.*;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
